Reorganised code to remove circular dependency
