+++ monitors/tests/test_health_check.py
@@ -8,7 +8,8 @@ import csv
 import os
 import mock
 
-from monitors.health_check import HealthCheckThread, write_last_run
+from monitors.health_check import HealthCheckThread
+from monitors.end_of_run_monitor import write_last_run
 from monitors.settings import EORM_LAST_RUN_FILE
 
 
@@ -48,26 +49,6 @@ class TestServiceUtils(unittest.TestCase):
         self.assertFalse(health_check_thread.exit)
         health_check_thread.stop()
         self.assertTrue(health_check_thread.exit)
-
-    def test_write_last_run(self):
-        """
-        Test write of the last runs CSV file
-        """
-        test_file = "last_runs_test.csv"
-        write_last_run(test_file, 'GEM', '1234')
-        write_last_run(test_file, 'WISH', '1234')
-        write_last_run(test_file, 'POLARIS', '1234')
-
-        row_array = [['GEM', '1234'], ['WISH', '1234'], ['POLARIS', '1234']]
-
-        # Now read back using the CSV reader
-        with open(test_file, 'r') as last_run_file:
-            last_run_reader = csv.reader(last_run_file)
-            for (i, row) in enumerate(last_run_reader):
-                self.assertEqual(row_array[i][0], row[0])
-                self.assertEqual(row_array[i][1], row[1])
-
-        os.remove(test_file)
 
     def test_thread_start_stop(self):
         health_check_thread = HealthCheckThread(1)
